+++ tests/AuthAPITestCase.py
@@ -1,7 +1,8 @@
 from rest_framework.test import APITestCase, APIClient
 from users.models import User
 from org.models import *
-
+from org.custom_model_field import PermissionSet as Permissions
+import uuid
 
 class AuthAPITestCase(APITestCase):
     """
@@ -18,6 +19,39 @@ class AuthAPITestCase(APITestCase):
         super(childClass,self).setUp() should be called for 
         a proper setUp of AuthHelper class.
         """
+        
+        """
+            Create an organization in the test database
+        """
+        # Create the org
+        self.org = Org.objects.create(
+            route_slug= 'route_slug',
+            name='test',
+            tagline='test',
+        )
+
+        # Create default group permissions  
+        self.volunteer_permissions = Permissions()
+        
+        #Create volunteer permission set objecct
+        self.volunteer_permission_set = PermissionSet.objects.create(
+            name = 'Volunteer',
+            org = self.org,
+            permissions = self.volunteer_permissions 
+        )
+
+
+        self.volunteer_group_invite_slug = str(self.org.id)+'-'+str(uuid.uuid4())
+
+        #Create volunteer group with organization
+        self.volunteer_group = Group.objects.create(
+            name='Volunteer',
+            role='''When a person clicks join org button of his own without a 
+                invite link then he will be put into this group.''',
+            invite_slug=self.volunteer_group_invite_slug,
+            org=self.org,
+            default_permission_set=self.volunteer_permission_set
+        )
 
         # create an un_authenticated user
         self.un_auth_user_email = "carbon.composite.dummy@gmail.com"
@@ -73,6 +107,25 @@ class AuthAPITestCase(APITestCase):
         client = APIClient()
         return client
 
+    #Creates auth client in the org as volunteer
+    def create_auth_client_in_org(self):
+        """
+        This function returns an authorized client.
+        """
+        login_api = "/api/users/login/"
+        login_payload = {
+            'email': self.auth_user_email,
+            'password': self.auth_user_password
+        }
+        add_volunteer_api = "/api/org/1/volunteer/"
+        auth_client = APIClient()
+        login_response = auth_client.post(login_api, login_payload)
+        auth_token = login_response.data['token']
+        
+        # Authorizing the requests by adding the token in header.
+        auth_client.credentials(HTTP_AUTHORIZATION=auth_token)
+        auth_client.get(add_volunteer_api)
+        return auth_client
 
     def tearDown(self):
         """
